+++ changelog.md
@@ -32,6 +32,7 @@ Date format is `MM/DD/YYYY`.
 - isLoggedIn() method to check if the user is logged in
 - getGroupRoles() method to get a list of roles for a group
 - getGroupRoleMembers() method to get a list of members for a group role
+- getGroupMembers() method to get a list of members for a group
 
 ### Fixed